Fixed warnings in x64
git-svn-id: http://miranda.googlecode.com/svn/trunk@9268 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/src/modules/netlib/netlibhttp.cpp
@@ -610,7 +610,7 @@ void NetlibHttpSetLastErrorUsingHttpResult(int result)
 
 char* gzip_decode(char *gzip_data, int *len_ptr, int window)
 {
-    size_t gzip_len = *len_ptr * 5;
+    int gzip_len = *len_ptr * 5;
     char* output_data = NULL;
 
     int gzip_err;
